Add createDeck function
createDeck() function is a nested for loop function. It will loop first
through the suits and then using nested in it for loop it will run
through the values of the cards (we have two arrays suits and values
through which we loop). A the end we push all newly created cards to the
deck and return the deck.